Bunch of small changes
Asteroid field adjusted lowered so the player can actually reach it.
Comets have now got a model instead of being a cube. Comets spawn twice as often.
Changed movement script to player script. Added recovery boolean to be used when getting hit.
When hit by asteroids or comets, the player is teleported to the centre hub and is put on a recovery timer, effectively going to lock them into the low-intensity for a bit.
Changed some collisions to require to search for player tag.